# apps/exposed-api/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Record_any.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ResponsePayload": {
				"properties": {
					"code": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.any_"
					},
					"message": {
						"type": "string"
					},
					"pagination": {
						"$ref": "#/components/schemas/Record_any.any_"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse": {
				"properties": {
					"resCode": {
						"type": "number",
						"format": "double"
					},
					"resData": {
						"$ref": "#/components/schemas/ResponsePayload"
					}
				},
				"required": [
					"resCode",
					"resData"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.ChallengeMode": {
				"type": "string",
				"enum": [
					"OneVsOne",
					"Group",
					"Partial"
				]
			},
			"ChallengeMode": {
				"$ref": "#/components/schemas/_36_Enums.ChallengeMode"
			},
...
			"_36_Enums.Outcome": {
				"type": "string",
				"enum": [
					"Win",
					"Lose",
					"CancelledOrDraw"
				]
			},
			"Outcome": {
				"$ref": "#/components/schemas/_36_Enums.Outcome"
			},
			"_36_Enums.CategoryDepth": {
				"type": "string",
				"enum": [
					"Pickem",
					"DayPickem",
					"WeekPickem",
					"Game",
					"Team",
					"Player"
				]
			},
			"CategoryDepth": {
				"$ref": "#/components/schemas/_36_Enums.CategoryDepth"
			},
			"_36_Enums.ParticipantRole": {
				"type": "string",
				"enum": [
					"Creator",
					"Participant",
					"Initiator",
					"Initializer"
				]
			},
			"ParticipantRole": {
				"$ref": "#/components/schemas/_36_Enums.ParticipantRole"
			},
...
			"JoinChallenge": {
				"properties": {
					"participantAccountId": {
						"type": "string"
					},
					"paidWalletAddress": {
						"type": "string"
					},
					"oddsFlag": {
						"type": "boolean"
					},
					"multiTokenFlag": {
						"type": "boolean"
					},
					"participantOutcome": {
						"$ref": "#/components/schemas/Outcome"
					},
					"participantOdds": {
						"type": "number",
						"format": "double"
					},
					"participationValueQty": {
						"type": "number",
						"format": "double"
					},
					"scChallengeId": {
						"type": "string"
					},
					"transactionHash": {
						"type": "string"
					},
					"challengeValueQty": {
						"type": "number",
						"format": "double"
					},
					"participationValueUsd": {
						"type": "number",
						"format": "double"
					},
					"exchangeRate": {
						"type": "number",
						"format": "double"
					},
					"challengeMode": {
						"$ref": "#/components/schemas/ChallengeMode"
					},
					"gameType": {
						"$ref": "#/components/schemas/CategoryDepth"
					},
					"contractId": {
						"type": "number",
						"format": "double"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"groupId": {
						"type": "number",
						"format": "double"
					},
					"subgroupId": {
						"type": "number",
						"format": "double"
					},
					"challengeGroupId": {
						"type": "number",
						"format": "double"
					},
					"participantStatP1": {
						"type": "number",
						"format": "double"
					},
					"participantStatP2": {
						"type": "number",
						"format": "double"
					},
					"statAttribute": {
						"type": "string"
					},
					"participantRole": {
						"$ref": "#/components/schemas/ParticipantRole"
					},
					"participantInputQty": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"participantAccountId",
					"paidWalletAddress",
					"oddsFlag",
					"multiTokenFlag",
					"participantOutcome",
					"participantOdds",
					"participationValueQty",
					"scChallengeId",
					"transactionHash",
					"challengeValueQty",
					"participationValueUsd",
					"exchangeRate",
					"challengeMode",
					"gameType",
					"contractId",
					"challengeGroupId",
					"participantRole"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.ShareStatus": {
				"type": "string",
				"enum": [
					"Cancelled",
					"Sent",
					"Converted",
					"Approved",
					"Rejected"
				]
			},
			"ShareStatus": {
				"$ref": "#/components/schemas/_36_Enums.ShareStatus"
			},
			"UpdateChallengeMode": {
				"properties": {
					"challengeMode": {
						"$ref": "#/components/schemas/ChallengeMode"
					},
					"shareStatus": {
						"$ref": "#/components/schemas/ShareStatus"
					},
					"creatorStakedQty": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"challengeMode",
					"shareStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "exposed-api",
		"version": "1.1.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/v1/sports/challenges/{challengeId}": {
			"post": {
				"operationId": "JoinChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JoinChallenge"
							}
						}
					}
				}
			}
		},
...
		"/v1/games/challenges/{inviteCode}": {
			"get": {
				"operationId": "GetGamesInChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "inviteCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/challenges/contracts": {
			"get": {
				"operationId": "GetContracts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/challenges/{inviteCode}/lineups/{challengeResultId}": {
			"get": {
				"operationId": "GetPickemChallengeLineups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "inviteCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "challengeResultId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/requests": {
			"get": {
				"operationId": "GetPartialBetEvents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/{userId}": {
			"get": {
				"operationId": "GetUserChallenges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/{challengeId}/requests": {
			"patch": {
				"operationId": "UpdateChallengeMode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Sports"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateChallengeMode"
							}
						}
					}
				}
			}
		},
		"/v1/challenges/{inviteCode}/participants": {
			"get": {
				"operationId": "GetChallengeParticipants",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "inviteCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/{inviteCode}/leaderboard/{challengeResultId}": {
			"get": {
				"operationId": "GetChallengeLeaderboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "inviteCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "challengeResultId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/leagues/{leagueId}": {
			"get": {
				"operationId": "GetPublicChallenges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "leagueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/categories/games/{gameId}": {
			"get": {
				"operationId": "GetCategoriesInGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/categories/{categoryId}/groups": {
			"get": {
				"operationId": "GetGroupsInCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/challenges/groups/{groupId}/subgroups": {
			"get": {
				"operationId": "GetSubgroupsInGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "groupId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Local development server"
		}
	]
}